refactor(storage): split FsStorage by domain
The single 633-line fs.rs interleaved the struct, on-disk shapes
(SettingsFile, StepLine), helpers, the trait impl, and free fns.
Split into one submodule per Storage domain (sessions, agents, mcp,
skills, config, scaffold). The single trait impl stays in fs/mod.rs
as thin delegation; each domain owns its data shape and helpers.
Move atomic_write into fs/mod.rs so all imports go through
super::atomic_write.
