--- crates/crabtalk/src/storage/fs.rs
@@ -1,633 +1,0 @@
-//! Filesystem-backed [`Storage`] implementation.
-//!
-//! Single struct holding all paths and state. Replaces the four
-//! separate Fs*Repo structs and the FsStorage composite.
-
-use anyhow::Result;
-use parking_lot::Mutex;
-use serde::{Deserialize, Serialize};
-use std::{
-    collections::{BTreeMap, HashMap, HashSet},
-    fs,
-    io::ErrorKind,
-    path::PathBuf,
-};
-use wcore::{
-    AgentConfig, AgentId, ConversationMeta, DaemonConfig, EventLine, McpServerConfig,
-    model::HistoryEntry,
-    storage::{
-        SessionHandle, SessionSnapshot, SessionSummary, Skill, Storage, validate_table_name,
-    },
-};
-
-use super::{atomic_write, default_crab};
-
-/// Header prepended to `local/settings.toml`. Tells humans not to edit
-/// the file while the daemon is running and signals the file is
-/// daemon-owned.
-const SETTINGS_HEADER: &str = "\
-# Managed by crabtalk daemon. Edits while the daemon is running are
-# overwritten on the next write. Edits while the daemon is stopped are
-# picked up on next reload.
-#
-# Source of truth for runtime-added MCPs and agents. Immutable
-# install-time configuration (providers, task pool) lives in
-# config.toml.
-
-";
-
-/// Filesystem persistence backend.
-pub struct FsStorage {
-    /// Config directory root (for agent prompt storage under `agents/<ulid>/`).
-    config_dir: PathBuf,
-    /// Root for session directories.
-    sessions_root: PathBuf,
-    /// Ordered skill roots to scan (local first, then packages).
-    skill_roots: Vec<PathBuf>,
-    /// Per-session step counters, recovered from disk on first access.
-    session_counters: Mutex<HashMap<String, u64>>,
-}
-
-impl FsStorage {
-    pub fn new(config_dir: PathBuf, sessions_root: PathBuf, skill_roots: Vec<PathBuf>) -> Self {
-        Self {
-            config_dir,
-            sessions_root,
-            skill_roots,
-            session_counters: Mutex::new(HashMap::new()),
-        }
-    }
-
-    // ── Session helpers ────────────────────────────────────────────
-
-    fn session_dir(&self, slug: &str) -> PathBuf {
-        self.sessions_root.join(slug)
-    }
-
-    fn session_meta_path(&self, slug: &str) -> PathBuf {
-        self.session_dir(slug).join("meta")
-    }
-
-    fn session_step_path(&self, slug: &str, step: u64) -> PathBuf {
-        self.session_dir(slug).join(format!("step-{step:06}"))
-    }
-
-    fn next_step(&self, slug: &str) -> u64 {
-        let mut counters = self.session_counters.lock();
-        let counter = counters
-            .entry(slug.to_owned())
-            .or_insert_with(|| recover_step_counter(&self.session_dir(slug)));
-        let n = *counter;
-        *counter += 1;
-        n
-    }
-
-    fn write_step(&self, slug: &str, line: StepLine) -> Result<()> {
-        let step = self.next_step(slug);
-        let path = self.session_step_path(slug, step);
-        if let Some(parent) = path.parent() {
-            fs::create_dir_all(parent)?;
-        }
-        let bytes = serde_json::to_vec(&line)?;
-        atomic_write(&path, &bytes)
-    }
-
-    // ── Agent helpers ──────────────────────────────────────────────
-
-    fn agent_prompt_path(&self, id: &AgentId) -> PathBuf {
-        self.config_dir
-            .join("agents")
-            .join(id.to_string())
-            .join("prompt.md")
-    }
-
-    fn read_agent_prompt(&self, id: &AgentId) -> Option<String> {
-        if id.is_nil() {
-            return None;
-        }
-        fs::read_to_string(self.agent_prompt_path(id)).ok()
-    }
-
-    // ── Settings helpers ───────────────────────────────────────────
-
-    fn settings_path(&self) -> PathBuf {
-        self.config_dir.join(wcore::paths::SETTINGS_FILE)
-    }
-
-    /// Read and parse the settings file. Re-parsed on every call —
-    /// settings are tiny and writes are rare, so a cache would be
-    /// premature. Don't add one without measuring.
-    fn read_settings(&self) -> Result<SettingsFile> {
-        let path = self.settings_path();
-        match fs::read_to_string(&path) {
-            Ok(content) => Ok(toml::from_str(&content)?),
-            Err(e) if e.kind() == ErrorKind::NotFound => Ok(SettingsFile::default()),
-            Err(e) => Err(e.into()),
-        }
-    }
-
-    fn write_settings(&self, file: &SettingsFile) -> Result<()> {
-        let path = self.settings_path();
-        if let Some(parent) = path.parent() {
-            fs::create_dir_all(parent)?;
-        }
-        let body = toml::to_string_pretty(file)?;
-        let mut content = String::with_capacity(SETTINGS_HEADER.len() + body.len());
-        content.push_str(SETTINGS_HEADER);
-        content.push_str(&body);
-        atomic_write(&path, content.as_bytes())
-    }
-}
-
-// ── Private on-disk shape ──────────────────────────────────────────
-
-/// On-disk shape of `local/settings.toml`. Holds runtime-added records:
-///   - `[mcps.<name>]` — MCP server registrations
-///   - `[agents.<name>]` — full agent definitions (model, members, …)
-#[derive(Debug, Default, Serialize, Deserialize)]
-struct SettingsFile {
-    #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
-    mcps: BTreeMap<String, McpServerConfig>,
-    #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
-    agents: BTreeMap<String, AgentConfig>,
-}
-
-#[derive(Serialize, Deserialize)]
-#[serde(untagged)]
-enum StepLine {
-    Compact {
-        /// Name of the `Archive`-kind entry in `memory` whose content
-        /// is the compacted prefix of the session up to this point.
-        archive_name: String,
-        archived_at: String,
-    },
-    /// Pre-Phase-5 compact marker that stored the summary inline.
-    /// Still recognized on read so older sessions keep their replay
-    /// boundary; the inline summary is no longer available, but the
-    /// boundary itself prevents stale pre-compact history from being
-    /// replayed.
-    LegacyCompact {
-        compact: String,
-        #[serde(default)]
-        title: String,
-        #[serde(default)]
-        archived_at: String,
-    },
-    Event(EventLine),
-    Entry(HistoryEntry),
-}
-
-impl StepLine {
-    fn is_compact_boundary(&self) -> bool {
-        matches!(self, Self::Compact { .. } | Self::LegacyCompact { .. })
-    }
-}
-
-// ── Storage implementation ─────────────────────────────────────────
-
-impl Storage for FsStorage {
-    // ── Skills ─────────────────────────────────────────────────────
-
-    fn list_skills(&self) -> Result<Vec<Skill>> {
-        let mut skills = Vec::new();
-        let mut seen = HashSet::new();
-        for root in &self.skill_roots {
-            if !root.exists() {
-                continue;
-            }
-            let entries = match fs::read_dir(root) {
-                Ok(e) => e,
-                Err(_) => continue,
-            };
-            for entry in entries {
-                let entry = match entry {
-                    Ok(e) => e,
-                    Err(_) => continue,
-                };
-                let path = entry.path();
-                if !path.is_dir() {
-                    continue;
-                }
-                let name = match path.file_name().and_then(|n| n.to_str()) {
-                    Some(n) if !n.starts_with('.') => n.to_owned(),
-                    _ => continue,
-                };
-                if seen.contains(&name) {
-                    continue;
-                }
-                let skill_path = path.join("SKILL.md");
-                if !skill_path.exists() {
-                    continue;
-                }
-                let content = match fs::read_to_string(&skill_path) {
-                    Ok(c) => c,
-                    Err(e) => {
-                        tracing::warn!("failed to read {}: {e}", skill_path.display());
-                        continue;
-                    }
-                };
-                match crate::hooks::skill::loader::parse_skill_md(&content) {
-                    Ok(skill) => {
-                        seen.insert(name);
-                        skills.push(skill);
-                    }
-                    Err(e) => tracing::warn!("failed to parse {}: {e}", skill_path.display()),
-                }
-            }
-        }
-        Ok(skills)
-    }
-
-    fn load_skill(&self, name: &str) -> Result<Option<Skill>> {
-        for root in &self.skill_roots {
-            let skill_path = root.join(name).join("SKILL.md");
-            if !skill_path.exists() {
-                continue;
-            }
-            let content = fs::read_to_string(&skill_path)?;
-            let skill = crate::hooks::skill::loader::parse_skill_md(&content)?;
-            return Ok(Some(skill));
-        }
-        Ok(None)
-    }
-
-    // ── Sessions ───────────────────────────────────────────────────
-
-    fn create_session(&self, agent: &str, created_by: &str) -> Result<SessionHandle> {
-        let agent_slug = wcore::sender_slug(agent);
-        let sender = wcore::sender_slug(created_by);
-        let prefix = format!("{agent_slug}_{sender}_");
-        let seq = next_session_seq(&self.sessions_root, &prefix);
-        let slug = format!("{agent_slug}_{sender}_{seq}");
-
-        let dir = self.session_dir(&slug);
-        fs::create_dir_all(&dir)?;
-
-        let meta = ConversationMeta {
-            agent: agent.to_owned(),
-            created_by: created_by.to_owned(),
-            created_at: chrono::Utc::now().to_rfc3339(),
-            title: String::new(),
-            uptime_secs: 0,
-            topic: None,
-        };
-        let meta_bytes = serde_json::to_vec(&meta)?;
-        atomic_write(&self.session_meta_path(&slug), &meta_bytes)?;
-        Ok(SessionHandle::new(slug))
-    }
-
-    fn find_latest_session(&self, agent: &str, created_by: &str) -> Result<Option<SessionHandle>> {
-        let agent_slug = wcore::sender_slug(agent);
-        let sender = wcore::sender_slug(created_by);
-        let prefix = format!("{agent_slug}_{sender}_");
-
-        if !self.sessions_root.exists() {
-            return Ok(None);
-        }
-
-        let mut best: Option<(u32, String)> = None;
-        for entry in fs::read_dir(&self.sessions_root)? {
-            let entry = entry?;
-            let name = entry.file_name();
-            let name = name.to_string_lossy();
-            if !name.starts_with(&prefix) || !entry.file_type()?.is_dir() {
-                continue;
-            }
-            let seq_str = &name[prefix.len()..];
-            if let Ok(seq) = seq_str.parse::<u32>()
-                && best.as_ref().is_none_or(|(b, _)| seq > *b)
-            {
-                best = Some((seq, name.to_string()));
-            }
-        }
-        Ok(best.map(|(_, slug)| SessionHandle::new(slug)))
-    }
-
-    fn load_session(&self, handle: &SessionHandle) -> Result<Option<SessionSnapshot>> {
-        let slug = handle.as_str();
-        let meta_path = self.session_meta_path(slug);
-        let meta_bytes = match fs::read(&meta_path) {
-            Ok(b) => b,
-            Err(e) if e.kind() == ErrorKind::NotFound => return Ok(None),
-            Err(e) => return Err(e.into()),
-        };
-        let meta: ConversationMeta = serde_json::from_slice(&meta_bytes)?;
-
-        let dir = self.session_dir(slug);
-        let mut step_files: Vec<_> = fs::read_dir(&dir)?
-            .filter_map(|e| e.ok())
-            .filter(|e| {
-                e.file_name()
-                    .to_str()
-                    .is_some_and(|n| n.starts_with("step-"))
-            })
-            .collect();
-        step_files.sort_by_key(|e| e.file_name());
-
-        let mut lines = Vec::with_capacity(step_files.len());
-        let mut last_compact_idx: Option<usize> = None;
-        for entry in &step_files {
-            let bytes = fs::read(entry.path())?;
-            match serde_json::from_slice::<StepLine>(&bytes) {
-                Ok(line) => {
-                    if line.is_compact_boundary() {
-                        last_compact_idx = Some(lines.len());
-                    }
-                    lines.push(line);
-                }
-                Err(e) => {
-                    tracing::warn!("skipping unparsable step {}: {e}", entry.path().display());
-                }
-            }
-        }
-
-        let start = last_compact_idx.unwrap_or(0);
-        let mut history = Vec::new();
-        let mut archive = None;
-        for (i, line) in lines[start..].iter().enumerate() {
-            match line {
-                StepLine::Compact { archive_name, .. } if i == 0 && last_compact_idx.is_some() => {
-                    archive = Some(archive_name.clone());
-                }
-                StepLine::Entry(entry) => history.push(entry.clone()),
-                StepLine::Event(_) | StepLine::Compact { .. } | StepLine::LegacyCompact { .. } => {}
-            }
-        }
-
-        Ok(Some(SessionSnapshot {
-            meta,
-            history,
-            archive,
-        }))
-    }
-
-    fn list_sessions(&self) -> Result<Vec<SessionSummary>> {
-        if !self.sessions_root.exists() {
-            return Ok(Vec::new());
-        }
-        let mut summaries = Vec::new();
-        for entry in fs::read_dir(&self.sessions_root)? {
-            let entry = entry?;
-            if !entry.file_type()?.is_dir() {
-                continue;
-            }
-            let slug = entry.file_name().to_string_lossy().to_string();
-            let meta_path = self.session_meta_path(&slug);
-            if let Ok(bytes) = fs::read(&meta_path)
-                && let Ok(meta) = serde_json::from_slice::<ConversationMeta>(&bytes)
-            {
-                summaries.push(SessionSummary {
-                    handle: SessionHandle::new(slug),
-                    meta,
-                });
-            }
-        }
-        Ok(summaries)
-    }
-
-    fn append_session_messages(
-        &self,
-        handle: &SessionHandle,
-        entries: &[HistoryEntry],
-    ) -> Result<()> {
-        for entry in entries {
-            self.write_step(handle.as_str(), StepLine::Entry(entry.clone()))?;
-        }
-        Ok(())
-    }
-
-    fn append_session_events(&self, handle: &SessionHandle, events: &[EventLine]) -> Result<()> {
-        for event in events {
-            self.write_step(handle.as_str(), StepLine::Event(event.clone()))?;
-        }
-        Ok(())
-    }
-
-    fn append_session_compact(&self, handle: &SessionHandle, archive_name: &str) -> Result<()> {
-        let line = StepLine::Compact {
-            archive_name: archive_name.to_owned(),
-            archived_at: chrono::Utc::now().to_rfc3339(),
-        };
-        self.write_step(handle.as_str(), line)
-    }
-
-    fn update_session_meta(&self, handle: &SessionHandle, meta: &ConversationMeta) -> Result<()> {
-        let path = self.session_meta_path(handle.as_str());
-        if let Some(parent) = path.parent() {
-            fs::create_dir_all(parent)?;
-        }
-        let bytes = serde_json::to_vec(meta)?;
-        atomic_write(&path, &bytes)
-    }
-
-    fn delete_session(&self, handle: &SessionHandle) -> Result<bool> {
-        let dir = self.session_dir(handle.as_str());
-        match fs::remove_dir_all(&dir) {
-            Ok(()) => Ok(true),
-            Err(e) if e.kind() == ErrorKind::NotFound => Ok(false),
-            Err(e) => Err(e.into()),
-        }
-    }
-
-    // ── Agents ─────────────────────────────────────────────────────
-
-    fn list_agents(&self) -> Result<Vec<AgentConfig>> {
-        let file = self.read_settings()?;
-        let mut out = Vec::with_capacity(file.agents.len());
-        for (name, mut cfg) in file.agents {
-            cfg.name = name;
-            cfg.system_prompt = self.read_agent_prompt(&cfg.id).unwrap_or_default();
-            out.push(cfg);
-        }
-        Ok(out)
-    }
-
-    fn load_agent(&self, id: &AgentId) -> Result<Option<AgentConfig>> {
-        if id.is_nil() {
-            return Ok(None);
-        }
-        let file = self.read_settings()?;
-        let Some((name, mut cfg)) = file.agents.into_iter().find(|(_, c)| c.id == *id) else {
-            return Ok(None);
-        };
-        cfg.name = name;
-        cfg.system_prompt = self.read_agent_prompt(id).unwrap_or_default();
-        Ok(Some(cfg))
-    }
-
-    fn load_agent_by_name(&self, name: &str) -> Result<Option<AgentConfig>> {
-        let file = self.read_settings()?;
-        let Some(mut cfg) = file.agents.get(name).cloned() else {
-            return Ok(None);
-        };
-        cfg.name = name.to_owned();
-        cfg.system_prompt = self.read_agent_prompt(&cfg.id).unwrap_or_default();
-        Ok(Some(cfg))
-    }
-
-    fn upsert_agent(&self, config: &AgentConfig, prompt: &str) -> Result<()> {
-        if config.id.is_nil() {
-            anyhow::bail!("cannot upsert agent with nil ID");
-        }
-        if config.name.is_empty() {
-            anyhow::bail!("cannot upsert agent with empty name");
-        }
-        validate_table_name("agent", &config.name)?;
-        let mut file = self.read_settings()?;
-        file.agents.insert(config.name.clone(), config.clone());
-        self.write_settings(&file)?;
-        let path = self.agent_prompt_path(&config.id);
-        if let Some(parent) = path.parent() {
-            fs::create_dir_all(parent)?;
-        }
-        atomic_write(&path, prompt.as_bytes())
-    }
-
-    fn delete_agent(&self, id: &AgentId) -> Result<bool> {
-        let mut file = self.read_settings()?;
-        let removed_name = file
-            .agents
-            .iter()
-            .find(|(_, c)| c.id == *id)
-            .map(|(n, _)| n.clone());
-        let settings_removed = removed_name.is_some();
-        if let Some(name) = removed_name {
-            file.agents.remove(&name);
-            self.write_settings(&file)?;
-        }
-        let dir = self.config_dir.join("agents").join(id.to_string());
-        let dir_removed = match fs::remove_dir_all(&dir) {
-            Ok(()) => true,
-            Err(e) if e.kind() == ErrorKind::NotFound => false,
-            Err(e) => return Err(e.into()),
-        };
-        Ok(dir_removed || settings_removed)
-    }
-
-    fn rename_agent(&self, id: &AgentId, new_name: &str) -> Result<bool> {
-        validate_table_name("agent", new_name)?;
-        let mut file = self.read_settings()?;
-        let old_name = file
-            .agents
-            .iter()
-            .find(|(_, c)| c.id == *id)
-            .map(|(n, _)| n.clone());
-        let Some(old_name) = old_name else {
-            return Ok(false);
-        };
-        if old_name == new_name {
-            return Ok(true);
-        }
-        if file.agents.contains_key(new_name) {
-            anyhow::bail!("agent '{new_name}' already exists");
-        }
-        let cfg = file.agents.remove(&old_name).expect("present above");
-        file.agents.insert(new_name.to_owned(), cfg);
-        self.write_settings(&file)?;
-        Ok(true)
-    }
-
-    fn load_config(&self) -> Result<DaemonConfig> {
-        let path = self.config_dir.join(wcore::paths::CONFIG_FILE);
-        if !path.exists() {
-            return Ok(DaemonConfig::default());
-        }
-        DaemonConfig::load(&path)
-    }
-
-    fn save_config(&self, config: &DaemonConfig) -> Result<()> {
-        let path = self.config_dir.join(wcore::paths::CONFIG_FILE);
-        let content = toml::to_string_pretty(config)?;
-        atomic_write(&path, content.as_bytes())
-    }
-
-    fn scaffold(&self) -> Result<()> {
-        fs::create_dir_all(&self.config_dir)?;
-        fs::create_dir_all(self.config_dir.join(wcore::paths::LOCAL_DIR))?;
-        fs::create_dir_all(self.config_dir.join(wcore::paths::SKILLS_DIR))?;
-        fs::create_dir_all(self.config_dir.join(wcore::paths::AGENTS_DIR))?;
-        fs::create_dir_all(&self.sessions_root)?;
-
-        let file = self.read_settings()?;
-        if file.agents.is_empty() {
-            let crab = default_crab();
-            let prompt = crab.system_prompt.clone();
-            self.upsert_agent(&crab, &prompt)?;
-        }
-        Ok(())
-    }
-
-    // ── MCP servers ────────────────────────────────────────────────
-
-    fn list_mcps(&self) -> Result<BTreeMap<String, McpServerConfig>> {
-        let mut file = self.read_settings()?;
-        // Fill in `name` from the key for entries hand-edited into the
-        // file without one. In-memory only — caller sees normalized
-        // values but the file on disk is left as-is.
-        for (name, cfg) in file.mcps.iter_mut() {
-            if cfg.name.is_empty() {
-                cfg.name = name.clone();
-            }
-        }
-        Ok(file.mcps)
-    }
-
-    fn load_mcp(&self, name: &str) -> Result<Option<McpServerConfig>> {
-        Ok(self.read_settings()?.mcps.remove(name).map(|mut cfg| {
-            if cfg.name.is_empty() {
-                cfg.name = name.to_owned();
-            }
-            cfg
-        }))
-    }
-
-    fn upsert_mcp(&self, config: &McpServerConfig) -> Result<()> {
-        validate_table_name("mcp", &config.name)?;
-        let mut file = self.read_settings()?;
-        file.mcps.insert(config.name.clone(), config.clone());
-        self.write_settings(&file)
-    }
-
-    fn delete_mcp(&self, name: &str) -> Result<bool> {
-        let mut file = self.read_settings()?;
-        let removed = file.mcps.remove(name).is_some();
-        if removed {
-            self.write_settings(&file)?;
-        }
-        Ok(removed)
-    }
-}
-
-// ── Free functions ─────────────────────────────────────────────────
-
-fn recover_step_counter(dir: &PathBuf) -> u64 {
-    let mut max = 0u64;
-    if let Ok(entries) = fs::read_dir(dir) {
-        for entry in entries.flatten() {
-            let name = entry.file_name();
-            let name = name.to_string_lossy();
-            if let Some(suffix) = name.strip_prefix("step-")
-                && let Ok(n) = suffix.parse::<u64>()
-            {
-                max = max.max(n);
-            }
-        }
-    }
-    max + 1
-}
-
-fn next_session_seq(root: &PathBuf, prefix: &str) -> u32 {
-    let mut max = 0u32;
-    if let Ok(entries) = fs::read_dir(root) {
-        for entry in entries.flatten() {
-            let name = entry.file_name();
-            let name = name.to_string_lossy();
-            if let Some(seq_str) = name.strip_prefix(prefix)
-                && let Ok(seq) = seq_str.parse::<u32>()
-            {
-                max = max.max(seq);
-            }
-        }
-    }
-    max + 1
-}
